requirements.txt
streamlit
python-dotenv
notion-client
langchain
langchain-google-genai
faiss-cpu
